HealthFit/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
crispy-bootstrap5==2024.10
distro==1.9.0
Django==5.1.2
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
openai==1.54.3
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
psycopg2-binary==2.9.10
google==3.0.0
beautifulsoup4==4.12.3
soupsieve==2.6
google-generativeai==0.8.4